No need to use size_t here
However, I should/could size_t for array indices. My arrays are so small
that it doesn't matter most of the time, but it's more correct
nonetheless.

+++ asciifield.c
@@ -21,7 +21,7 @@ struct screen
 
     double speed;
     double fps;
-    size_t num_stars;
+    int num_stars;
     int first;
 
     int draw_ship;
@@ -229,7 +229,7 @@ void
 ensure_stars(struct screen *s, struct star **field)
 {
     struct star *p, *last = NULL;
-    size_t existing;
+    int existing;
 
     if (*field == NULL)
     {
